+++ TileMapEditor/Source/Interactables/BoxyBox.h
@@ -13,7 +13,7 @@ public:
     void setColor(sf::Color color);
 
     sf::RectangleShape getRect() { return rect; };
-    bool contains(float x, float y) const;
+    bool contains(int x, int y) const;
 
 private:
